Changeset:
* Simplify Makefile
* Accept input file as command line argument
* Increase path length to 1024 for all path-related variables
* Replace hard-coded path to \*.dat files with environment variable MOOG_DATA
* SM needs double precision for several of its functions. Pass REAL8 not REAL4 to those.

+++ Findtic.f
@@ -5,7 +5,7 @@ c     this routine makes tic marks at nice round intervals.
 c******************************************************************************
 
       implicit real*8(a-h,o-z)
-      real*4 end1, end2, bigtic, smltic, span, spanlog
+      real*8 end1, end2, bigtic, smltic, span, spanlog
 
 
       span = end2 - end1
